pyproject.toml
[project]
name = "healthisure"
version = "0.1.0"
description = "Healthisure: An OpenEnv RL environment for health insurance support resolution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
web = [
    "gradio",
    "pandas",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[dependency-groups]
dev = [
    "pytest",
    "httpx",
]
